Two updates to how test results are passed to SBT.
First, record the name of the test method as the test name instead of the name of the
test class. The test name is much more useful in the output than the test class.

Second, we will no longer pass the Throwable from TestNG's ITestResult if the test
succeeded. It's possible for a successful test to have a non-null Throwable, if for
example the test annotation specifies an expectedExceptions parameter. This change
stops printing those exceptions as errors.